customer-data-analytics/telegram-bot/requirements.txt
# Telegram Bot Dependencies
# Telegram Bot API
python-telegram-bot>=20.0

# HTTP client
httpx>=0.25.0
requests>=2.31.0

# Database connectivity
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Configuration and utilities
python-dotenv>=1.0.0
loguru>=0.7.0
pydantic>=2.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# API integration
../api

